refactored KeywordRadioBtnGroup component by constructing it from individual RadioBtn components
